Add select filter options Refactored code Added common css constants Added styles
