+++ pages/_app.tsx
@@ -1,4 +1,5 @@
 import "@/styles/globals.css";
+import "@/styles/variables.css";
 import type { AppProps } from "next/app";
 
 export default function App({ Component, pageProps }: AppProps) {
